Closes #302
Our polyfill for pointerdown and pointerup event handlers maps to touchstart and touchend events in certain browsers. The clickEatingDiv used by _Overlay hides itself when its click event handler fires.
We call preventDefault() in the pointerdown and pointerup event handlers, which works fine inside of the pointerdown and pointerup events, but when called inside of the touchstart and/or touchend events, preventDefault will stop a click from firing.
The solution is to detect when we are in a touchend event and call the click event handling function manually.